pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carry-regime-compass"
version = "0.1.0"
description = "Cross-asset carry-to-volatility monitor with macro regime inference."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "yfinance>=0.2.51",
    "pandas>=2.2,<3.0",
    "numpy>=1.26",
    "scipy>=1.13",
    "streamlit>=1.36",
    "plotly>=5.22",
    "streamlit-autorefresh>=1.0.1",
    "loguru>=0.7",
    "pydantic>=2.7",
    "PyYAML>=6.0",
    "tenacity>=9.0",
    "SQLAlchemy>=2.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3.14",
    "pytest-cov>=5",
    "freezegun>=1.5",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
carry-compass = "carry_compass.viz.app:main"

[tool.setuptools.packages.find]
include = ["carry_compass*"]

[tool.setuptools.package-data]
carry_compass = ["config/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
